Move instrumentation to fuzzilli.cpp
