+++ lib/VM/JSLib/GlobalObject.cpp
@@ -256,19 +256,6 @@ throwTypeError(void *ctx, Runtime *runtime, NativeArgs) {
   return runtime->raiseTypeError(TypeErrorMessage[kind]);
 }
 
-#ifdef HERMES_ENABLE_FUZZILLI
-CallResult<HermesValue> fuzzilli(void *, Runtime *, NativeArgs args) {
-  if (args.getArgCount() == 1 && args.getArg(0).isString()) {
-    if (args.getArg(0).getString()->equals("FuzzilliCrash1")){
-      *((int*) 0x1) = 2;
-    }
-    else if (args.getArg(0).getString()->equals("FuzzilliCrash2"))
-      assert(0);
-  }
-  return HermesValue::encodeUndefinedValue();
-}
-#endif
-
 // NOTE: when declaring more global symbols, don't forget to update
 // "Libhermes.h".
 void initGlobalObject(Runtime *runtime, const JSLibFlags &jsLibFlags) {
@@ -776,10 +763,6 @@ void initGlobalObject(Runtime *runtime, const JSLibFlags &jsLibFlags) {
         normalDPF,
         intl::createIntlObject(runtime)));
   }
-#endif
-
-#ifdef HERMES_ENABLE_FUZZILLI
-  defineGlobalFunc(Predefined::getSymbolID(Predefined::fuzzilli), fuzzilli, 1);
 #endif
 
 }
