+++ tools/CMakeLists.txt
@@ -14,8 +14,7 @@ add_subdirectory(hbc-deltaprep)
 add_subdirectory(hbc-attribute)
 add_subdirectory(jsi)
 add_subdirectory(emhermesc)
-add_subdirectory(libfuzzer)
-add_subdirectory(fuzzilli)
+add_subdirectory(fuzzers)
 add_subdirectory(dependency-extractor)
 add_subdirectory(hermes-parser)
