+++ tools/fuzzers/libfuzzer/fuzzer-jsi-entry.cpp
